Update index.html
added <meta charset=utf-8> so that the html supports encoding for all characters in unicode. For example, you can easily write — – “” without having to use &ndash; &mdash; etc. 

Also added ... , initial-scale=1"> to keep the initial scale of the viewport so that there aren't any issues with zooming on responsive/mobile